requirements.txt
colorcet==3.1.0
cycler==0.11.0
joblib==1.2.0
matplotlib==3.5.3
numba==0.57.0
numpy==1.24.1
numpy==1.22.3
pandas==1.5.3
scikit_allel==1.3.5
scipy==1.7.3
seaborn==0.13.2
sympy==1.12
tqdm==4.65.0
